+++ src/lib.rs
@@ -244,6 +244,7 @@ pub async fn set_shutdown(shutdown_sender: tokio::sync::watch::Sender<()>) -> an
     shutdown_sender.send(())?;
     Ok(())
 }
+
 /// Handles database related execution, using `H` for domain-specific database
 /// queries. Will attempt to reconnect to database on errors.
 async fn use_db<D, P, H>(
@@ -386,7 +387,7 @@ where
     H: NodeHooks<D>, {
     // Use TLS if the URI scheme is HTTPS.
     // This uses whatever system certificates have been installed as trusted roots.
-    let node_ep = if node_ep.uri().scheme() == Some(&http::uri::Scheme::HTTPS) {
+    let node_ep = if node_ep.uri().scheme() == Some(&concordium_rust_sdk::v2::Scheme::HTTPS) {
         node_ep.tls_config(ClientTlsConfig::new()).map_err(NodeError::ConnectionError)?
     } else {
         node_ep